QNA database/requirements.txt
psycopg2-binary==2.9.9
google-generativeai==0.7.2
python-dotenv==1.0.0